Add base class to modals
